+++ src/mctpd.c
@@ -1536,7 +1536,7 @@ static int peer_set_mtu(struct ctx *ctx, struct peer *peer, uint32_t mtu) {
 		return -EPROTO;
 	}
 
-	rc = mctp_nl_route_del(ctx->nl_query, peer->eid, ifname, NULL);
+	rc = mctp_nl_route_del(ctx->nl_query, peer->eid, 0, ifname, NULL);
 	if (rc < 0 && rc != -ENOENT) {
 		warnx("%s, Failed removing existing route for eid %d %s",
 			__func__,
@@ -1544,7 +1544,7 @@ static int peer_set_mtu(struct ctx *ctx, struct peer *peer, uint32_t mtu) {
 		// Continue regardless, route_add will likely fail with EEXIST
 	}
 
-	rc = mctp_nl_route_add(ctx->nl_query, peer->eid, ifname, NULL, mtu);
+	rc = mctp_nl_route_add(ctx->nl_query, peer->eid, 0, ifname, NULL, mtu);
 	if (rc >= 0) {
 		peer->mtu = mtu;
 	}
@@ -2243,10 +2243,10 @@ static int peer_route_update(struct peer *peer, uint16_t type)
 	}
 
 	if (type == RTM_NEWROUTE) {
-		return mctp_nl_route_add(peer->ctx->nl_query, peer->eid,
+		return mctp_nl_route_add(peer->ctx->nl_query, peer->eid, 0,
 					 link, NULL, peer->mtu);
 	} else if (type == RTM_DELROUTE) {
-		return mctp_nl_route_del(peer->ctx->nl_query, peer->eid,
+		return mctp_nl_route_del(peer->ctx->nl_query, peer->eid, 0,
 					 link, NULL);
 	}